Refactor context code placement
