+++ src/components/MarketSelector/GmPoolsSelectorForGlvMarket.tsx
@@ -14,6 +14,7 @@ import { useGlvGmMarketsWithComposition } from "components/Synthetics/MarketStat
 import { ButtonRowScrollFadeContainer } from "components/TableScrollFade/TableScrollFade";
 import TokenIcon from "components/TokenIcon/TokenIcon";
 
+import { useTokensFavorites } from "context/TokensFavoritesContext/TokensFavoritesContextProvider";
 import {
   GlvInfo,
   MarketInfo,
@@ -24,7 +25,6 @@ import {
 } from "domain/synthetics/markets";
 import { isGlvInfo } from "domain/synthetics/markets/glv";
 import { convertToUsd } from "domain/synthetics/tokens";
-import { useTokensFavorites } from "domain/synthetics/tokens/useTokensFavorites";
 import { stripBlacklistedWords } from "domain/tokens/utils";
 
 import Modal from "../Modal/Modal";
