+++ src/components/Sorter/Sorter.tsx
@@ -1,8 +1,8 @@
 import cx from "classnames";
 import React, { PropsWithChildren, useCallback, useRef } from "react";
 
-import { useSorterConfig } from "./SorterProvider";
-import { SortDirection, SorterConfig, SorterKey } from "./sorterTypes";
+import { useSorterConfig } from "context/SorterContext/SorterContextProvider";
+import { SortDirection, SorterConfig, SorterKey } from "context/SorterContext/types";
 
 import IcSortable from "img/ic_sortable.svg?react";
 import IcSortedAsc from "img/ic_sorted_asc.svg?react";
