+++ src/components/Synthetics/GmList/sortGmTokensByField.tsx
@@ -2,11 +2,11 @@ import values from "lodash/values";
 
 import { getMarketListingDate } from "config/markets";
 
+import type { SortDirection } from "context/SorterContext/types";
 import { MarketTokensAPRData, MarketsInfoData, getMintableMarketTokens } from "domain/synthetics/markets";
 import { getIsBaseApyReadyToBeShown } from "domain/synthetics/markets/getIsBaseApyReadyToBeShown";
 import { convertToUsd, type TokensData } from "domain/synthetics/tokens";
 
-import type { SortDirection } from "components/Sorter/sorterTypes";
 import type { SortField } from "./GmList";
 import { sortGmTokensDefault } from "./sortGmTokensDefault";
